+++ android/app/src/main/java/app/proxora/ProxoraApp.kt
@@ -1,0 +1,5 @@
+package app.proxora
+
+import android.app.Application
+
+class ProxoraApp : Application()